Add `--config` command for pointing to the config file
By default, pyzet should look in user home directory, so the pyzet
command can detect config from any place by default. Alternative is to
put pyzet config in zet repo, but in this case it wouldn't be able to
detect it automatically from any place.

Also, if not config is found, pyzet should be able to load defaults. But
then it woiuld also work only from the zet repo itself (defaults cannot
know where is the user's zet folder).

This is on very early stage at this point, still to be implemented.

Also in this commit:
- add new pre-commit hooks
- fix mypy hook, so it is able to install additional type stubs