Add command that runs `git status` in zet repo
In the future, it might be easier to add a full git interface, so all
git commands can be supported. The alternative is to add only most
needed commands (or we can add both ;])

Flake8 max-complexity needed to be increased again. I need to check
if this option is even used in bigger projects.