Allow for custom git options in `pyzet status`
Argparse tries to detect any option that starts with `-`, but we want to
trasfer these to the git. In order to do this, a user needs to insert
`--` before providing options like: `git status -s`.

So, the example call to the above from pyzet would look like this:

    pyzet status -- -s

It's not required for non-minus options, but it's easier to tell the
user to use it every time - it won't break anything.

Ref:
https://docs.python.org/3/library/argparse.html#arguments-containing