Add `_commit_zettel` and commit when editing zettel
`ED:` prefix is added to the commit message when zettel is edited. I
think that it will be quite helpful in the future for backtracking.

The zettel title should be limited to 50 characters, so it is not
snipped in GitHub, and here we are adding 4 more characters. IMO, the
information that zettel was editted is worth more than these four
potential characters of the title.