Add `--reverse` flag to `list` and make oldest first
When you have some zettels it's nice to have newest at the bottom, so
they are visible in the terminal. `--reverse` can be helpful if you want
to stream output to file etc.

So now, the behavior is similar to `tags`, where the most used tag is at
the bottom by default.