Fix the default value for passing options to git
I used a list with `"."` as the only element, and git seemed not to like
it in push command, as it didn't want to actually push.

The value was fixed to be an empty list instead. It seems to work fine,
IDK why I haven't used it in the first place.