Check if file is in git before adding `ED:` to msg
The current version always adds `ED:` as prefix to commit message if
only `pyzet edit` command is used. This is not the ideal way, because
the zettel can be not yet committed (and only a file exists), e.g.
because pre-commit rejected it.

In this situation, the user will use edit command to fix the issue, and
teh commit message should be the standard one.

Close #6.