Abort adding zettel if title is empty of whitespace only
This allows the user to remove everything from the zettel to abort,
adding it, running pre-commit etc. The user can leave some whitespace
because it's not visible, so it's good idea to also ignore any of it.

New test was added that checks if `get_markdown_title()` raises
`ValueError` if stripped title is equal to empty string. This plays nice
with remaining code where we also raise ValueError, and act upon it,
e.g. in `add_zettel()`.