Add different variant of grep test for Windows/Unix
For some reason, the output is ordered differently when comparing grep
from Git Bash on Windows to grep from Linux. Ultimately, this should be
fixed to have the same output on both platforms, but this isn't possible
within grep, and other command (like `sort`) or postprocessing in Python
is needed to make it work.