Add quick start tutorial
The most basic interaction with pyzet is explained there which can be
treated as a tutorial.

The link is global, so it will work with PyPI, but it also means that it
will work only after this file is a part of the main branch.

Close #12.

Also, remove TODO list from the readme. I switched to just use GitHub
issues with that. Easier to automate with `Close #<number>` syntax.