Add `pyzet init` command that creates a git repo
Close #8.

Also, update `get_zettels()`, so it displays an error message if ZETDIR
dir doesn't exist. Before that, we had a nasty error message. It was
found, because `pyzet init` wasn't creating a `zettels` folder
initially. Now it does, but this new error message is still with
us, if user happens to remove this folder from any reason.