Add support for YAML config file
It now supports four fields: repo, editor, git, and grep. Only the first
one is required to be present. Hard-coded default values are used when
other fields are not defined.

Executable paths defined in repo are treated as `str` which is converted
to `pathlib.Path` just before calling them. The actual looking for
executable is done with the help of `shutil.which()` that is able to
also look for it on `PATH`.

Other changes from this commit:

* Add `sample-config` command

  A config file can be easily created by redirecting stdout to a file.

    I took the inspiration here from pre-commit:
    https://github.com/pre-commit/pre-commit/blob/master/pre_commit/commands/sample_config.py

* Update `Config` dataclass:

    * Remove default values. Use get method to check if data is in YAML.
      If not, use the default value (apart from repo path that always
      needs to be set).

    * Rename `repo_path` to `repo` in Config.

* Go back to the single test variant for grep, and mark it as skipped,
  because grep behaves differently in Linux and Git Bash on Windows.

Also:

* Add error message where there is no config file present
* Add error message when `pyzet edit` is called when there are no zettels
* Add tests for already existing `init` command
* Update summary of commands in readme

Close #13.