Add `--link` to `show` as a way to obtain a MD link
It's another take on #21, but this one enables to create a link from a
single zettel, given their ID is already a part of `pyzet show` command.

Still, existing approach with `pyzet list --link` might be good when we
want to obtain a list of links, that e.g. all include a certain keyword
(which can be achieved when piping it with grep).

Update changelog.