Run `pyzet grep` from ZETDIR to shorten output lines
By default, `git grep` that is used here will print the path relative to
the folder from which it was run. So, running it from `zettels` folder
will reduce some clutter from the output which is good in 80 char
terminals.

More context in #23 which needs some more changes to be closed.