Add tox env for measuring coverage
With previous settings, coverage report was created with paths going to
`.tox` folder which are longer, and which were incompatible with
Coverage Gutters extension for VS Code.

Also, remove tox from requirements-dev.txt, because up to this point tox
was always installed by tox which doesn't make sense and only slows the
process down.

Update readme to account for this update.