Use `print()` for CLI output and logging for reporting
According to:
https://docs.python.org/3/howto/logging.html#when-to-use-logging

`print()` should be used to display console output for ordinary usage of
a CLI program.

Logging level was changed to WARNING, so info and debug messages will be
hidden from the user.

Add some logging in various functions to get a better idea what is going
on when executing standard commands.

Update log formatting, so it is `LEVEL: <log message>`.

Also, update pre-commit hooks