Add `pyzet remote` command
It's a simple interface to `git remote` similarly to some other
Git-interaction commands.

Also, update readme and docs, especially regarding Git interaction
commands which now should be described in a more understandable way with
some real examples.

The tutorial was reorganized a bit, and new subsections were added for
easier orientation.

Also, remove printing the output for the user in init_repo() because Git
displays a meaningful output on its own, so there is no sense in
duplication.

Close #41.