Freeze `Config`, use `evolve()` to create a new object
This is currently used only in special case with `pyzet init` when the
config value should be ignored, and the one given by the user is used
instead to initialize a new repo in any location.

Alternatively, `pyzet init` command could be changed to allow only for
passing repo parameter.