Stop unit testing on macOS
The behavior inside the tests is slightly different, and I don't have
a mac for proper debugging. Currently, the code should be working, and
the only behavior might be due to the differences in 'git grep', at
aleast according to the failing unit test which happens since commit:

    6a2b7ab22413f758b3d7207ef361fdc2e99f905e

It's strange, as I didn't have any specific settings for macOS.
I suppose that even if the previous version worked, it might worked only
because GitHub actions have very specific Git setup.