[WIP] Completely redesign the control flow of the app
pyzet starts to be more stateful, and it asks what to do next, after
previous action (like adding, opening, printing) was finished.

This is a huge change in the philosophy how to use the tool, but I
really like the new flow.

This is still big WIP, there are issues with a lot of tests, and very
likely a lot of hidden bugs which I will be gradually fixing with time.

I still believe that this is worth the effort, as working with the new
flow is very convenient in my ZK repo with 500+ notes.