Fix mocking of `builtins.input`, so tests can finish
It seems that `side_effect` has to be used over `return_value`.
`KeyboardInterrupt` seems to work pretty well as the 2nd side effect.

I'm not sure if this is the best appraoch, but these tests were broken
last time I modified the code which was 9 months ago, so for now I just
want to fix them.

Note that 3 tests still fail at this point.