Add a new `pyzet query` behaviors which checks only against titles
This is something new in `pyzet`, as so far we're looking always also at
the contents of the notes. Query is for high-level searching based on
keywords which are used in titles.

It replaces the previous `query` command which was a wrapper around an
infinite loop behavior which is a bit problematic, and would be
refactored in the future.