Resign from the infinite loop idea
It turned out that this idea is not so great, and it makes code much
less friendly to be expanded with AppState being a big bag for global
variables.

I'll be refactoring it, as there are some cool leftovers like getting
the zettel from grep with the whole logic around it.