Fix test_sample_error on MacOS
I noticed that this tests fails locally when trying to run on my
Macbook on Python 3.13. I checked in the pipeline after reincluding
MacOS, and 3.12 also started failing, so it behaves differently which is
weird.

Anyway, it turns out that sometimes quotes are skipped in list of
allowed options, i.e. it looks like: `(choose from unix, windows)`.

Now test checks for substring in the error which works in both cases.